bin2c: add the -p option for specifying path stripping
The goal here is to support out-of-tree builds of software. When this
happens, the source filename(s) given to bin2c may include additional
path components, such as "../images/foo.png", and this will
unwantedly influence the emitted variable name. Besides, it won't
look good either: bin2c____images_foo_png.

Add an option, -p, with which the number of paths to strip to the
left can be specified, similar to /usr/bin/patch's -p. To accomodate
arbitrary path prefixes, -p can also take negative numbers which
specify the number of components to the right to keep.

	bin2c -p0 i/foo.png     => bin2c_i_foo_png
	bin2c -p1 ../i/foo.png  => bin2c_i_foo_png
	bin2c -p-2 ../i/foo.png => bin2c_i_foo_png

A Makefile that supports out-of-tree builds would therefore always
use bin2c with a negative number.